+++ pkg/ui/src/Sidebar/SidebarNav.svelte
