+++ pkg/ui/src/Sidebar/index.ts
@@ -1,3 +1,3 @@
 export { default as SidebarItem } from './SidebarItem.svelte';
 export { default as SidebarItemGroup } from './SidebarItemGroup.svelte';
-export { default as SideBar } from './Sidebar.svelte';
+export { default as SideBarNav } from './SidebarNavNav.svelte';
